+++ crates/cli-support/src/wit/standard.rs
@@ -91,6 +91,7 @@ pub enum AdapterType {
     Struct(String),
     Enum(String),
     StringEnum(String),
+    DynamicUnion(String),
     NamedExternref(String),
     Function,
     NonNull,