+++ crates/cli/tests/reference/web-sys.d.ts
@@ -6,7 +6,7 @@
  * *This API requires the following crate features to be activated: `MediaSourceEnum`*
  */
 
-type MediaSourceEnum = "camera" | "screen" | "application" | "window" | "browser" | "microphone" | "audioCapture" | "other";
+export type MediaSourceEnum = "camera" | "screen" | "application" | "window" | "browser" | "microphone" | "audioCapture" | "other";
 
 export function get_media_source(): MediaSourceEnum;